layout, footer, privat route - login, offer page
