+++ src/authorizationStatus.ts
@@ -1,0 +1,4 @@
+import { AuthorizationStatus } from './const';
+
+export const getAuthorizationStatus = () => AuthorizationStatus.Auth;
+
