+++ src/components/logo-main-page.tsx
@@ -1,0 +1,7 @@
+export default function LogoMainPage(): JSX.Element {
+  return (
+    <a className="header__logo-link header__logo-link--active">
+      <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+    </a>
+  );
+}
