+++ src/components/logo.tsx
@@ -1,8 +1,9 @@
 import {Link} from 'react-router-dom';
+import { AppRoute } from '../const';
 
 export default function Logo(): JSX.Element {
   return (
-    <Link className="header__logo-link header__logo-link--active" to="/">
+    <Link className="header__logo-link" to={AppRoute.Root}>
       <span className="visually-hidden">На главную страницу</span>
       <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
     </Link>
